fix: comprehensive multi-tenant security and data integrity audit (39 rounds)
Security fixes:
- Remove 8 debug telemetry fetch() calls to localhost:7826 in Terminal.tsx
- Fix SQL injection via unquoted heredoc in instalar.sh (now uses psql -v)
- Move install_token from URL query params to Authorization header
- Add chmod 600 for .env, titan-agent.json, INSTALL_SUMMARY.txt
- Add webPreferences (contextIsolation, sandbox) to owner-app Electron
- Fix PIN timing attack: constant-time iteration over all users
- Reject JWT tokens missing jti claim with 401
- Add server-side token revocation (serverLogout in frontend)
- Stop exposing internal str(e) to HTTP clients in cfdi_service/tunnel
- Set entrypoints to set -euo pipefail

Data integrity fixes:
- Fix double stock deduction on merma approval (shrinkage_tracker)
- Systematic float→Decimal conversion across 15+ modules
- Fix REAL/DOUBLE PRECISION→NUMERIC(12,2) in 6 fiscal table schemas
- Remove float() wrappers on Decimal values in fiscal/routes.py (6 sites)
- Fix employees routes writing 0.0 float to NUMERIC columns
- Fix dashboard IVA arithmetic using float literal 1.16→Decimal
- Add migration 048 for fiscal NUMERIC column type changes

Validation hardening:
- Fix Pydantic str=None annotation bugs (customer_name, sale_id)
- Add ge=1 to 14+ entity ID fields across schemas
- Add min_length=1 to sku fields in products/remote schemas
- Add max_length to unbounded notes/hash_id fields
- Add ISO8601 pattern to SyncPushPayload.timestamp
- Add Literal pattern to notification_type
- Change BatchVarianceItem extra="allow"→"forbid"
- Escape LIKE wildcards in SAT catalog search

Other fixes:
- Fix migrations 033/044/045/046 missing schema_version INSERT
- Move bcrypt PIN verification outside transaction (reduce lock time)
- Fix infinite re-render in Terminal.tsx (useState for config)
- Fix timezone stripping in sync routes
- Add Docker log rotation to compose files
- Make CP schema_migrations creation unconditional
- revoke_token now returns bool instead of None

+++ backend/modules/fiscal/intercompany_billing.py
@@ -135,7 +135,7 @@ class CrossBranchBilling:
                 original_rfc TEXT,
                 target_rfc TEXT,
                 cross_concept TEXT,
-                amount DOUBLE PRECISION,
+                amount NUMERIC(12,2),
                 timestamp TIMESTAMP
             )
         """)
